Frameworks: add ability to send a custom intent after keyguard is finished
We have and will have some features which need to start intents after the user
has sucessfully entered the credentials. Add for the system public available method
to start an avtivity after the keyguard.

Add as well support on slim Action for it so that our features actually can use this.

Change-Id: I65b32dab7986f8d71a3a94597324412847e4e138

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/CommandQueue.java
	packages/SystemUI/src/com/android/systemui/statusbar/tv/TvStatusBar.java

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/CommandQueue.java