Fix IME layout breakage introduced in 54e79624fff3fb232fb1d4dc9dc9dc5350207eca
If we've brought up the keyboard and we have the additional ime layout tools enabled
we want to see them.  Under strange circumstances where there's no back or home key
present it's possible to get "stuck" in this layout with the keyboard up.  The
previous patchset modified the workaround for this to just not do anything in this
circumstance. Simply ignoring the situation is a bad way to handle things because
it makes the ime layout option appear broken and the user is expecting to see a
change.  Also, layout changes to the left were broken after the change in the constant
value and lets keep the private getters inlined, kthnx.

Change-Id: Ia51fa58f3b14831485fa6abdc2e551320efabf0d

Conflicts:
	core/java/com/android/internal/util/vanir/NavbarConstants.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/NavigationBarView.java