[1/2] Legacy Menu Customizations
First, thanks to slimroms for making it possible with their navbar development.

This adds the ability to set custom single and long actions to both left and right small menu buttons.
It supports custom app/shortcut as well.
Also if you do not like the spacing on the edges when legacy menu buttons are disabled, you can disable the layout and have the navigation bar completly without spaces.

Any improvement of code is welcome, it is my first feature so it may have some "mistakes" but everything works like a charm.

Developed by: @siracuervo - Axxion-Team ROMs
Thanks and mentions: SlimRoms, TeamEOS, Exodus, CM.

Change-Id: I25b0883b4b49343554d5f4b5103adb861fa7e74b