Frameworks: Add back app-info to notification longpress
This adds a new button in notification guts, which executes the app info activity, going to the notification app info.
Also, changes the notification inspect drawable to the "ringer_audible" one that makes more sense with the icon porpouse. I approached the implementation and made the notification inspect button long-clickeabl, which executes notification log.

Change-Id: Ic349e4d2ab1d15c2a01f5c229612c7929a30caba